requirements.txt
# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3

# Vector Database
chromadb==0.4.22

# Embeddings
sentence-transformers==2.3.1

# LLM - Groq (FREE)
groq==0.9.0
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1

# Development
flake8==7.0.0
pytest==7.4.3